Moved out configuration load/store from ui_driver.c to new ui_configuration.c More moves to come.
