+++ mchf-eclipse/usb/usbd/class/audio/inc/usbd_audio_core.h
@@ -136,7 +136,7 @@ typedef struct UsbAudioUnit_s {
 	int16_t max;
 	uint16_t res;
 	int16_t cur;
-	uint8_t* ptr; // pointer to data structure which is used elsewhere to store volume
+	__IO uint8_t* ptr; // pointer to data structure which is used elsewhere to store volume
 } UsbAudioUnit;
 
 // In / Out Volume;
